Refactor media handling by moving media-related code to quickSettingsMedia.js and updating integration in quickSettingsNotifications.js and extension.js
