+++ stylesheet.css
@@ -327,21 +327,38 @@ Gjs_apps_windowControls_WindowControlsIndicator.panel-button:focus {
 
 .message-list-clear-button {
   margin: 4px 8px;
-  padding: 4px;
-  border-radius: 6px;
-  background-color: transparent;
-  border: none;
+  /* Fixed base size: 16x16 */
   width: 16px;
   height: 16px;
+  padding: 0;
+  border-radius: 4px; /* small rounded corners for a squared look */
+  background-color: transparent;
+  border: none;
   -st-icon-style: symbolic;
+  transition: background-color 150ms ease, box-shadow 150ms ease;
+}
+
+.message-list-clear-button:focus {
+  box-shadow: inset 0 0 0 2px alpha(currentColor, 0.25);
+}
+
+.message-list-clear-button.disabled,
+.message-list-clear-button.disabled:hover,
+.message-list-clear-button.disabled:active {
+  background-color: transparent;
+  box-shadow: none;
+  opacity: 0.6;
 }
 
 .message-list-clear-button:hover {
+  /* Apply hover to the button background (not just the icon) */
   background-color: alpha(currentColor, 0.15);
+  box-shadow: inset 0 0 0 1px alpha(currentColor, 0.20);
 }
 
 .message-list-clear-button:active {
-  background-color: alpha(currentColor, 0.2);
+  background-color: alpha(currentColor, 0.20);
+  box-shadow: inset 0 0 0 1px alpha(currentColor, 0.25);
 }
 
 /* Notification close buttons - ensure circular and same size as expand */
